cleaned files and changed directory structures
