+++ host/host_configs/config2.json
@@ -1,0 +1,54 @@
+[
+  {
+    "vm_cid": 35,
+    "vm_name": "charmander",
+    "pcpu": 8,
+    "workload_config": {
+      "max_threads": 200,
+      "slices": [
+        {
+          "type": "repeater",
+          "cnt": 15,
+          "slices": [
+            {
+              "type": "time_slice",
+              "percentage_load": 0.2,
+              "interval": 5
+            },
+            {
+              "type": "time_slice",
+              "percentage_load": 0.4,
+              "interval": 5
+            }
+          ]
+        }
+      ]
+    }
+  },
+  {
+    "vm_cid": 36,
+    "vm_name": "charmander2",
+    "pcpu": 8,
+    "workload_config": {
+      "max_threads": 200,
+      "slices": [
+        {
+          "type": "repeater",
+          "cnt": 15,
+          "slices": [
+            {
+              "type": "time_slice",
+              "percentage_load": 0.2,
+              "interval": 5
+            },
+            {
+              "type": "time_slice",
+              "percentage_load": 0.4,
+              "interval": 5
+            }
+          ]
+        }
+      ]
+    }
+  }
+]
